# bilemo_api/public/swagger/swagger.yaml
swagger: '2.0'
info:
  title: Bilemo Api
  version: 1.0.1
  servers:
    - url: http://127.0.0.1:8000
paths:
  /api/products/{page}:
    get:
      tags:
        - Products
      summary: Get All Products
      parameters:
        - name: page
          in: path
          required: true
          type: integer
      responses:
        '200':
          description: Returns all products
        '404':
          description: Products not found
  '/api/product/{id}':
    get:
      tags:
        - Products
      summary: Get One product
      parameters:
        - name: id
          in: path
          required: true
          type: integer
      responses:
        '200':
          description: Returns one product
        '404':
          description: Product not found
  /api/users/{page}:
    get:
      tags:
        - User
      summary: Get All Users
      parameters:
        - name: page
          in: path
          required: true
          type: integer
      responses:
        '200':
          description: Returns all users
        '404':
          description: No user found
  '/api/users':
    post:
      tags:
        - User
      summary: Create an user
      consumes:
        - application/json
      parameters:
        - in: body
          name: user
          description: The user to create.
          schema:
            type: object
            required:
              - userName
            properties:
              firstname:
                type: string
              lastname:
                type: string
              password:
                type: string
              email:
                type: string
      responses:
        '201':
          description: Create an user
        '400':
          description: Bad request
  '/api/user/{id}':
    get:
      tags:
        - User
      summary: Get One user
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        '200':
          description: Returns one user
        '404':
          description: User not found
    delete:
      tags:
        - User
      summary: Delete an user
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        '204':
          description: Removes an user
        '400':
          description: Bad Request
host: 127.0.0.1:8000